+++ inc/functions/query.php
@@ -182,7 +182,11 @@ function set_rules_query( $query ) {
  * @return bool
  */
 function setup_post( $post_id = null ) {
-	global $post;
+	global $post, $content_control_last_post;
+
+	if ( ! isset( $content_control_last_post ) ) {
+		$content_control_last_post = [];
+	}
 
 	if ( is_null( $post_id ) ) {
 		return false;
@@ -195,6 +199,7 @@ function setup_post( $post_id = null ) {
 		( is_int( $post_id ) && $post_id !== $current_post_id );
 
 	if ( $overload_post ) {
+		$content_control_last_post[] = isset( $post ) ? $post : $current_post_id;
 		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 		$post = get_post( $post_id );
 		setup_postdata( $post );
@@ -206,13 +211,16 @@ function setup_post( $post_id = null ) {
 /**
  * Check and clear global post if needed.
  *
- * @param bool $overload_post Whether post was overloaded.
- *
  * @return void
  */
-function clear_post( $overload_post = false ) {
-	if ( $overload_post ) {
+function clear_post() {
+	global $post, $content_control_last_post;
+	if ( ! empty( $content_control_last_post ) ) {
+		// Get the last post ID from the array.
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
+		$post = array_pop( $content_control_last_post );
+
 		// Reset global post object.
-		wp_reset_postdata();
+		setup_postdata( $post );
 	}
 }