+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+## v2.0.9 - 09/24/2023
+
+-   Improvement: Better handling of restriction titles & content. Admins with priv can insert any content into the restriction messages.
+-   Improvement: Added new filter `content_control/query_filter_init_hook` to allow delaying query filtering for compatibility with plugins that make custom queries before `template_redirect` action.
+
+```php
+add_filter( 'content_control/query_filter_init_hook', function () {
+    return 'init'; // Try setup_theme, after_theme_setup, init or wp_loaded
+} );
+```
+
+-   Tweak: Ensure our restriction checks work within a nested post loop.
+-   Tweak: Change how restriction title & descriptions were sent/received over Rest API.
+-   Fix: Bug that caused some shortcodes to not render properly.
+-   Fix: Bug where override message wasn't used.
+-   Fix: Bug where Elementor Post loop would render incorrectly when using ACF fields in the loop template.
+
 ## v2.0.8 - 09/22/2023
 
 -   Tweak: Ignore many Elementor queries from being restricted.